requirements.txt
Adafruit_DHT-Fixed~=1.4.2
pygame~=2.0.1
termcolor~=1.1.0